Set up to run as a CLI
