+++ bin/interpreter/error.js
