+++ bin/interpreter/index.js
