+++ bin/interpreter/symbolTable.js
