+++ bin/interpreter/types/boolean.js
