+++ bin/interpreter/types/number.js
