+++ bin/interpreter/types/string.js
