+++ bin/lexer/index.js
