+++ bin/lexer/position.js
