+++ bin/lexer/token.js
