+++ bin/lexer/tokenTypes.js
