+++ bin/parser/parseResult.js
