+++ bin/utils/stringWithArrows.js
